+++ backend/app.js
@@ -1,0 +1,18 @@
+const express = require('express');
+const cors = require('cors');
+const app = express();
+
+app.use(cors());
+// app.use(bodyParser.json());
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+// routes
+app.get('/', (req, res) => {
+    res.send('API is running 🚀');
+});
+
+app.use('/signup', require('./src/routes/user/signup.js'));
+app.use('/login', require('./src/routes/user/login.js'));
+
+module.exports = app
